requirements.txt
boto3==1.17.89
PyYAML==5.4.1